+++ auto-matching-api-rails/lib/auto_matching/converter/converter_base.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class ConverterBase
+    end
+  end
+end
