+++ auto-matching-api-rails/lib/auto_matching/converter/happymail.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class Happymail < ConverterBase
+    end
+  end
+end
