+++ auto-matching-api-rails/lib/auto_matching/converter/ikukuru.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class Ikukuru < ConverterBase
+    end
+  end
+end
