+++ auto-matching-api-rails/lib/auto_matching/converter/merupara.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class Merupara < ConverterBase
+    end
+  end
+end
