+++ auto-matching-api-rails/lib/auto_matching/converter/mint.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class Mint < ConverterBase
+    end
+  end
+end
