+++ auto-matching-api-rails/lib/auto_matching/converter/wakuwaku.rb
@@ -1,0 +1,6 @@
+module AutoMatching
+  module Converter
+    class Wakuwaku < ConverterBase
+    end
+  end
+end
