+++ auto-matching-api-rails/spec/lib/auto_matching/converter/pcmax_converter_spec.rb
@@ -1,0 +1,5 @@
+require "spec_helper"
+
+describe "Dog" do
+  it "is named 'Pochi'"
+end
